Turtle class now takes doubles not floats
Uses an implcit conversion for processing calls that need floats

+++ src/test/scala/TurtleSpec.scala
@@ -11,7 +11,7 @@ class TurtleSpec extends Specification {
 
   def createNewTurtle(): Turtle = {
     val pa = mock(classOf[PApplet])
-    new Turtle(0.0f, 0.0f, 0.0f, pa)
+    new Turtle(0.0, 0.0, 0.0, pa)
   }
 
   "The Turtle class" should {
@@ -21,29 +21,29 @@ class TurtleSpec extends Specification {
     }
     "convert degrees to radians correctly" in {
       val t = createNewTurtle()
-      t.deg2Rad(90.0f) must_== (Pi/2).toFloat
+      t.deg2Rad(90.0) must_== (Pi/2)
     }
 
     "return a new Turtle instance when it's moved" in {
       val t = createNewTurtle()
-      val newT = t.move(10.0f)
+      val newT = t.move(10.0)
       newT must haveClass[Turtle]
     }
     "update the position of the new Turtle instance" in {
       val t = createNewTurtle()
-      val newT = t.move(10.0f)
-      newT.y must_== 10.0f
+      val newT = t.move(10.0)
+      newT.y must_== 10.0
     }
 
     "return a new Turtle instance when it's turned" in {
       val t = createNewTurtle()
-      val newT = t.turn(10.0f)
+      val newT = t.turn(10.0)
       newT must haveClass[Turtle]
     }
     "update the heading of the new Turtle instance" in {
       val t = createNewTurtle()
-      val newT = t.turn(10.0f)
-      val newHeading = t.deg2Rad(10.0f)
+      val newT = t.turn(10.0)
+      val newHeading = t.deg2Rad(10.0)
       newT.heading must_== newHeading
     }
